[AOT Refactor] traced function transforms pt. 2
[ghstack-poisoned]
